requirements.txt
aiogram~=3.15.0
python-dotenv~=1.0.1
fitz~=0.0.1.dev2
easyocr~=1.7.2
aiofiles~=24.1.0
pillow~=10.4.0
openai~=1.57.4
phidata
googletrans-py
sqlalchemy
pgvector
psycopg2